move new tests files to appropriate folders
